awscosts/requirements.txt
boto3==1.28.9
psycopg2_binary==2.9.7
